测试。

+++ app/src/com/twirlingvr/www/utils/DownloadReceiver.java
@@ -6,6 +6,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.net.Uri;
+import android.text.TextUtils;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -35,14 +36,17 @@ public class DownloadReceiver extends BroadcastReceiver {
             if (item == null) {
                 return;
             }
-            String fileFolder = item.getAndroidoffline().substring(0, item.getAndroidoffline().length() - 4);
-            if (item.getIsatmos() == 1) {
-                new Decompress(Constants.PAPH_DOWNLOAD_LOCAL + item.getAndroidoffline(), Constants.PAPH_DOWNLOAD_LOCAL + fileFolder).unzip();
-            }
+            String fileFolder = "";
             RealmHelper.getIns().updateDownloadId(id);
+            if (!TextUtils.isEmpty(item.getAndroidoffline())) {
+                fileFolder = item.getAndroidoffline().substring(0, item.getAndroidoffline().length() - 4);
+                if (item.getIsatmos() == 1) {
+                    new Decompress(Constants.PAPH_DOWNLOAD_LOCAL + item.getAndroidoffline(), Constants.PAPH_DOWNLOAD_LOCAL + fileFolder).unzip();
+                }
+                FileUtil.delete(new File(Constants.PAPH_DOWNLOAD_LOCAL + item.getAndroidoffline()));
+                FileUtil.delete(new File(Constants.PAPH_DOWNLOAD_LOCAL + fileFolder));
+            }
             printDownloadInformation(id);
-            FileUtil.delete(new File(Constants.PAPH_DOWNLOAD_LOCAL + item.getAndroidoffline()));
-            FileUtil.delete(new File(Constants.PAPH_DOWNLOAD_LOCAL + fileFolder));
             //TODO 判断这个id与之前的id是否相等，如果相等说明是之前的那个要下载的文件
 //            if (id == App.downloadId) {
 //                Intent install = new Intent(Intent.ACTION_VIEW);
